requirements.txt
streamlit>=1.31,<2
pandas>=2.1
numpy>=1.26
plotly>=5.18
fastapi>=0.108
uvicorn>=0.24
requests>=2.31
httpx>=0.24,<0.28
python-multipart>=0.0.6
openpyxl>=3.1
matplotlib>=3.8
seaborn>=0.13
python-dotenv>=1.0
colorama>=0.4.6
pydantic>=2.5
typing-extensions>=4.9
pyarrow>=14 